improve Windows command prompt template
Mappings for the Windows command prompt mapped the bright colours to
shades of light/dark, which resulted in weird looking applications.
This new mapping respects the bright colours, and mostly uses the same
colours for the dark and light versions, because the colour schemes do
not have dark and light versions of the colours.

The mapping is optimised for the Solarized scheme. For example, 0F in
Solarized maps perfectly to the purple/pink of cmd, but 0F in other
schemes looks more brown. Brown-ish colours could be more suitable for
the dark red colour of cmd, but Solarized 0F would look stange when
used as dark red. I think this mapping is a good compromise.